(refactor) preparing code for fastify migration
