+++ src/services/exerciseGroups/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './api';
+export * from './hooks';
