+++ src/services/exerciseGroups/types.ts
@@ -1,0 +1,7 @@
+export interface ExerciseGroup {
+  id: string;
+  name: string;
+  color: string;
+  num_exercises_to_show: number;
+  user_id: string;
+}
